Add RPC support to the compiler and stub rpc dsl for extensibility
Even though the original thought behind protobuf-core was to stip
rpc from the project, it doesn't make sense to have the core
compiler leave out support for services (part of the protobuf spec).

By stubbing out the RPC dsl, we can still lift the actual RPC
implementation to a different (and potentially custom) gem, while
keeping the protobuf compliant with the spec.